feat: auth ready for codereview

+++ internal/auth/claims.go
@@ -3,9 +3,9 @@ package auth
 import "github.com/golang-jwt/jwt/v5"
 
 type Claims struct {
-	UserID    int    `json:"user_id"`
-	Username  string `json:"username"`
-	Role      string `json:"role"`
-	EstateIDs []int  `json:"estate_ids"`
+	UserID    int64   `json:"user_id"`
+	Username  string  `json:"username"`
+	Role      string  `json:"role"`
+	EstateIDs []int64 `json:"estate_ids"`
 	jwt.RegisteredClaims
 }
